% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_snapshot_data.R
\name{expect_snapshot_data}
\alias{expect_snapshot_data}
\title{Snapshot a data frame}
\usage{
expect_snapshot_data(x, name, digits = 6)
}
\arguments{
\item{x}{a \link{data.frame}}

\item{name}{a \link{character} name of the snapshot}

\item{digits}{a \link{numeric} number of digits to round to}
}
\description{
Snapshot a data frame
}
\examples{
\dontrun{
iris |> expect_snapshot_data(name = "iris")
}
}
