% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.data-package.R
\docType{package}
\name{snapshot.data-package}
\alias{snapshot.data}
\alias{snapshot.data-package}
\title{snapshot.data: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/snapshot.data}
  \item Report bugs at \url{https://github.com/d-morrison/snapshot.data/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
